Improve and simplify macOS builds (#1276)
* Add missing macOS requirement for 'make package'

We need packages for the last build step on macOS, update docs
to reflect the requirement.

* Add build instructions using homebrew

On macOS, we can use a package manager to easily install
dependencies. This simplifies onboarding and building Veracrypt.

* Add flag to use homebrew packages

When building, we can use prebuilt wxwidgets from homebrew to
simplify and speed up local building. We also put the package
behind a flag as it's optional during development.

* Skip signing for local builds

When building with homebrew, skip signing. This can be put behind
a flag to enable, if needed.

* Use system yasm on macOS if available

The binary in the repo is not universal (x86_64) and therefore
building fails on arm architecture if Rosetta is not installed.

Use local yasm if available.

* Build local arch only in development

When building via homebrew and locally, build only the local arch
which skips ASM for arm(Mx) for MacOS. This removes the need to
have rosetta installed for building.

* Fix compilation issue when COMPILE_ASM is undefined

Use a conditional check for COMPILE_ASM not being false instead of true.
This avoids passing the variable to other parts of the build script.

* Set SDK 12 as the minimum requirement and target

Align the requirement to SDK 12 in both the makefile and script,
and update the comment to remove confusion.
I chose to leave this on 12 to be on the side of err and support
as many building platforms as possible, when we can support.

The local script now also sets the target using the local sdk
version. This should improve the local development experience.

* Fix wrong architecture for macOS in x86 builds

We now build only the current arch for local development builds
in macOS. This change also fixes the x86 builds failing.

* Add instructions brew backed macOS local builds

Flags to build a local build using homebrew packages are not
default and require parameter -b to build. We also don't build
packages directly, which requires -p.

* Fix wxwidgets not linking in local x86 macOS development builds

* Clarify build location in the document